+++ MASTER_S3-P4/P4_JC1060P470C/src/display/UIPage3B.h
@@ -1,8 +1,0 @@
-// src/display/UIPage3B.h
-#pragma once
-#include "lvgl.h"
-
-void uiPage3BCreate();
-void uiPage3BUpdate();
-void uiPage3BDestroy();
-void uiPage3BCreate(lv_obj_t* parent);